fix: address retrieval quality review comments
Keep eval runtime failures fatal, tighten explain metadata tests, prioritize supplemental insertion mode, and split retrieval tests under the monolith limit.

Resolves review thread PRRT_kwDORS2O8s6CgsI9

Resolves review thread PRRT_kwDORS2O8s6CgsJB

Resolves review thread PRRT_kwDORS2O8s6CgsJF

Resolves review thread PRRT_kwDORS2O8s6Cgq2C

Resolves review thread PRRT_kwDORS2O8s6CgsJD

Resolves review thread PRRT_kwDORS2O8s6CgrNS

Resolves review thread PRRT_kwDORS2O8s6CgrNT

Resolves review thread PRRT_kwDORS2O8s6CgrNV
